requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
spacy>=3.7.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
